requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl
frozenlist==1.4.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.0
openai==0.28.1
packaging==23.2
pathy==0.10.2
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
yarl==1.9.2
